bluetooth: add BLE HOGP stack, vhid transport, and HCI LE fixes
Add Bluetooth Low Energy HID device support via HOGP (HID over GATT
Profile) with a userspace daemon and kernel virtual HID transport.

Kernel changes:
- ng_hci: handle LE_Read_Buffer_Size to use dedicated LE buffer pool,
  fix le_connection_complete_ep.latency type (u_int8_t -> u_int16_t),
  add Enhanced Connection Complete (0x0a) handler for LE Privacy,
  handle LTK Request events, fix LE buffer accounting in
  num_compl_pkts and send_data_packets
- vhid: new virtual HID transport kernel module exposing /dev/vhid
  control device and /dev/vhidN instances, implements hid_if KOBJ
  interface so hidbus/hkbd/hms/hmt attach automatically

Userspace (usr.sbin/bluetooth/btled):
- ATT protocol client (Core Spec Vol 3 Part F)
- SMP LE Legacy Pairing with Just Works (Core Spec Vol 3 Part H)
- GATT service/characteristic/descriptor discovery (Core Spec Vol 3 Part G)
- HOGP profile: HID Service discovery, Report Map, report notifications
- HCI utilities: BLE scanning, adapter address, connection handle lookup
- Capsicum sandbox after initialization

Prototype status — not yet validated on hardware.

+++ sys/netgraph/bluetooth/include/ng_hci.h
@@ -2117,7 +2117,7 @@ typedef struct {
 	u_int8_t 	address_type;
 	bdaddr_t	address;
 	u_int16_t 	interval;
-	u_int8_t	latency;
+	u_int16_t	latency;
 	u_int16_t	supervision_timeout;
 	u_int8_t	master_clock_accuracy;
